Move allocation tracing to MemoryBank
